backend/fix/international-fleet-bugs

+++ Backend/dev/clickhouse/sql-seed/atlas-driver-offer-bpp-seed.sql
@@ -158,10 +158,10 @@ create table atlas_driver_offer_bpp.fleet_rc_daily_stats (
     `fleet_owner_id` String,
     `rc_id` String,
     `merchant_local_date` Date DEFAULT today(),
-    `total_completed_rides` Int64,
-    `total_earnings` Int64,
-    `ride_distance` Int64,
-    `ride_duration` Int64,
+    `total_completed_rides` Int32,
+    `total_earnings` Int32,
+    `ride_distance` Int32,
+    `ride_duration` Int32,
     `version` DateTime DEFAULT now()
 ) ENGINE = ReplacingMergeTree(version)
 ORDER BY (merchant_local_date, fleet_owner_id, rc_id);
